+++ example/lib/main.dart
@@ -16,13 +16,16 @@ Future<void> main() async {
   await BackgroundLocationTrackerManager.initialize(
     _backgroundCallback,
     config: const BackgroundLocationTrackerConfig(
+      loggingEnabled: true,
       androidConfig: AndroidConfig(
         notificationIcon: 'explore',
         trackingInterval: Duration(seconds: 4),
+        distanceFilterMeters: 5,
       ),
       iOSConfig: IOSConfig(
         activityType: ActivityType.FITNESS,
         distanceFilterMeters: 5,
+        restartAfterKill: true,
       ),
     ),
   );