+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## 1.1.0 - 12-10-2021
+## 1.1.0 - xx-xx-2021
 #Added
 - Added more options to android config to specify update interval
 - Added iOS specific options to control activity type and/or distance filter
+- Added iOS specific options to restart the tracking after killing the app
+- Added option to Android specific options to only get updates every x meters
 
 ## 1.0.2 - 22-06-2021
 #Fixed